refactor(clean-architecture): Split into submodules
