+++ app/src/main/java/com/github/kereis/medit/editor/BlockTextAction.kt
@@ -1,0 +1,19 @@
+package com.github.kereis.medit.editor
+
+import android.widget.EditText
+
+class BlockTextAction(private val token: String) : TextAction() {
+    private val startBlock = "\n$token"
+    private val endBlock = "\n\n$token\n"
+
+    override fun apply(editTextView: EditText) {
+        val end = editTextView.selectionEnd
+
+        val content = editTextView.text
+
+        if (!editTextView.hasSelection()) {
+            content.insert(end, "$startBlock$endBlock")
+            editTextView.setSelection(end + startBlock.length)
+        }
+    }
+}
