+++ app/src/main/java/com/github/kereis/medit/format/markdown/BoldTextAction.kt
@@ -1,0 +1,11 @@
+package com.github.kereis.medit.format.markdown
+
+/**
+ * **BoldTextAction**
+ *
+ * Applies bold text formatting at current selection.
+
+ * @see InlineSpanTextAction
+ * @version 1.0.0
+ */
+// class BoldTextAction : InlineSpanTextAction("**")
