+++ app/src/main/java/com/github/kereis/medit/parser/MarkdownParser.kt
@@ -1,0 +1,27 @@
+package com.github.kereis.medit.parser
+
+import android.content.Context
+import androidx.fragment.app.Fragment
+import io.noties.markwon.Markwon
+import com.github.kereis.medit.ui.editor.MarkdownRenderFragment
+
+sealed class MarkdownParser {
+    companion object {
+        /**
+         * This method returns a `android.view.View` with parsed Markdown text.
+         * It can be used to exchange fragments when render is requested.
+         *
+         * @return `androidx.fragment.app.Fragment` containing the Markdown text
+         */
+        fun render(context: Context, input: String): Fragment {
+            val parser = makeParser(context)
+            val parsedInput = parser.parse(input)
+            val renderedSpanned = parser.render(parsedInput)
+            return MarkdownRenderFragment(renderedSpanned)
+        }
+
+        private fun makeParser(context: Context): Markwon {
+            return Markwon.create(context)
+        }
+    }
+}
