+++ app/src/main/java/com/github/kereis/medit/ui/explorer/FileExplorerStorageListFragment.kt
@@ -1,0 +1,7 @@
+package com.github.kereis.medit.ui.explorer
+
+import androidx.fragment.app.Fragment
+import com.github.kereis.medit.R
+
+class FileExplorerStorageListFragment : Fragment(R.layout.fragment_file_explorer_storage_list) {
+}
