--- app/src/main/java/me/humenius/medit/format/markdown/BoldTextAction.kt
@@ -1,14 +1,0 @@
-package me.humenius.medit.format.markdown
-
-import me.humenius.medit.editor.InlineSpanTextAction
-
-/**
- * **BoldTextAction**
- *
- * Applies bold text formatting at current selection.
- *
- * @author humenius
- * @see InlineSpanTextAction
- * @version 1.0.0
- */
-// class BoldTextAction : InlineSpanTextAction("**")
