--- app/src/main/java/me/humenius/medit/ui/editor/TextActionButton.kt
@@ -1,14 +1,0 @@
-package me.humenius.medit.ui.editor
-
-import android.content.Context
-import android.util.AttributeSet
-
-class TextActionButton : androidx.appcompat.widget.AppCompatImageButton {
-    constructor(context: Context) : super(context)
-    constructor(context: Context, attrs: AttributeSet?) : super(context, attrs)
-    constructor(context: Context, attrs: AttributeSet?, defStyleAttr: Int) : super(
-        context,
-        attrs,
-        defStyleAttr
-    )
-}
