--- app/src/main/java/me/humenius/medit/ui/explorer/FileExplorerStorageListFragment.kt
@@ -1,7 +1,0 @@
-package me.humenius.medit.ui.explorer
-
-import androidx.fragment.app.Fragment
-import me.humenius.medit.R
-
-class FileExplorerStorageListFragment : Fragment(R.layout.fragment_file_explorer_storage_list) {
-}
