--- app/src/test/java/me/humenius/medit/ExampleUnitTest.kt
@@ -1,21 +1,0 @@
-package me.humenius.medit
-
-import org.junit.Assert
-import org.junit.Test
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * See [testing documentation](http://d.android.com/tools/testing).
- */
-class ExampleUnitTest {
-    @Test
-    fun addition_isCorrect() {
-        Assert.assertEquals(4, 2 + 2)
-    }
-
-    @Test
-    fun subtraction_isCorrect() {
-        Assert.assertEquals(0, 2 - 2)
-    }
-}
