+++ medit-abstractcode/src/main/java/com/github/kereis/medit/abstractcode/MyClass.kt
@@ -1,0 +1,4 @@
+package com.github.kereis.medit.abstractcode
+
+class MyClass {
+}
