+++ medit-adapters/src/main/java/com/github/kereis/medit/adapters/MyClass.kt
@@ -1,0 +1,4 @@
+package com.github.kereis.medit.adapters
+
+class MyClass {
+}
