+++ settings.gradle
@@ -1,2 +1,8 @@
+include ':test123'
+include ':medit-abstractcode'
+include ':medit-plugins'
+include ':medit-adapters'
+include ':medit-domain'
 include ':app'
+
 rootProject.name = "medit"
